rename to xsynth; add "special" names
XSynth is a pun on synthx.exe, a tool in XACTstep (ds371 to be precise, ie.
part of the ABEL / AHDL toolchain) which does basically the same thing:
transform BLIF to XNF. however, synthx is rather picky about the input it
accepts (no comments, only XNF-valid names) and then crashes with a null
pointer dereference anyway. the only complicated part that synthx does is
implementing BLIF state machines, but those are more easily coded in Verilog
anyway.